move auth code to separate source file, add support for enabling auth using env var
